Mise à jour de l'écouteur Changement de l'aireDeJeu (Je sais pas si c'est bien mais ça marche bien) Debut de la gestion du tour dans moteur Rajout de la zone d'attaque dans Pion

+++ trunk/moteur/Noeud.java
@@ -12,10 +12,10 @@ import java.util.ArrayList;
  */
 public class Noeud {
     public Case c;
-    public int cout;
+    public double cout;
     public ArrayList<Noeud> listeNoeud;
     
-    public Noeud(Case c, int cout){
+    public Noeud(Case c, double cout){
         this.c = c;
         this.cout = cout;
         listeNoeud = new ArrayList<Noeud>();
